Re-introduced "TV" prefix on all interfaces
Interfaces needs a prefix for namespacing not to confuse developers and avoid
possible name collisions with other specs. Note the group may decide to use
another prefix than "TV" in the end.

The update also adds missing inheritance to EventTarget for the interfaces
that were recently introduced and that use events.