Spec converted to WebIDL contiguous mode (automatic conversion)
The conversion hardly changes anything from a visual perspective. Under the
hoods, the WebIDL definitions that used `<dl>` have been replaced by more
readable `<pre>` definitions (which ReSpec calls the WebIDL contiguous mode).

I used a tool to convert the spec and tidied the result so that it remains
easy to read and maintain.

Note the conversion tool generated the "attributes" and "methods" `<dl>` lists
which now have to be manually edited. This may prove to be a pain to maintain:
the attribute types and method parameters tables duplicate the info in the
WebIDL definitions and are easy to get wrong. On top of that, the descriptions
of the methods would best be linked to the procedures for these methods, and
that's not the case today. We may switch to a "prose-oriented" approach as in
other specs later on, though.

Side note that there were 2 attributes named `name` in the `TVChannel`
interface definition. I renamed the one of type `TVChannelAttributes` to
`attributes`.